Editing excluir methods to throw an exception when the id is not defined in the database.

+++ src/main/java/com/algaworks/algafoodapi/domain/service/CadastroCozinhaService.java
@@ -21,12 +21,14 @@ public class CadastroCozinhaService {
     }
 
     public void excluir(Long cozinhaId) {
+        if (cozinhaRepository.findById(cozinhaId).isEmpty()) {
+            throw new EntidadeNaoEncontradaException(
+                    String.format("Não existe um cadastro de cozinha com o código %d.", cozinhaId));
+        }
+
         try {
             cozinhaRepository.deleteById(cozinhaId);
 
-        } catch (EmptyResultDataAccessException e) {
-            throw new EntidadeNaoEncontradaException(
-                    String.format("Não existe um cadastro de cozinha com o código %d", cozinhaId));
         } catch (DataIntegrityViolationException e) {
             throw new EntidadeEmUsoException(
                     String.format("Cozinha de código %d não pode ser removida, pois está em uso", cozinhaId));
